+++ fili-core/src/test/groovy/com/yahoo/bard/webservice/web/DataApiRequestFilterSpec.groovy
@@ -62,7 +62,7 @@ class DataApiRequestFilterSpec extends Specification {
         tg.getDimensions() >> dimensionDict.apiNameToDimension.values()
         table = new LogicalTable("name", DAY, tg)
         dimensionDict.apiNameToDimension.values().each {
-            DimensionColumn.addNewDimensionColumn(table, it, new PhysicalTable("abc", DAY.buildZonedTimeGrain(DateTimeZone.UTC), [:]))
+            DimensionColumn.addNewDimensionColumn(table, it)
         }
     }